backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
alembic>=1.13.1
psycopg2-binary>=2.9.9

# ── Validation & settings ─────────────────────────────────────────────────────
pydantic[email]>=2.7.1
pydantic-settings>=2.3.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.x reads bcrypt.__about__; bcrypt 4.1+ removed it , pin until auth uses bcrypt directly
bcrypt>=4.0.1,<4.1

# ── File uploads ──────────────────────────────────────────────────────────────
python-multipart>=0.0.9

# ── PDF extraction ────────────────────────────────────────────────────────────
pdfplumber>=0.11.0

# ── LangChain & LLM ──────────────────────────────────────────────────────────
langchain>=0.2.5
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.5
openai>=1.33.0

# ── Vector stores ─────────────────────────────────────────────────────────────
chromadb>=0.5.3
pinecone>=3.2.2

# ── Hybrid retrieval ──────────────────────────────────────────────────────────
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# ── Observability ─────────────────────────────────────────────────────────────
langsmith>=0.1.77
